# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: knowpod-service API
  version: 0.7.0
  description: |
    Backend for the knowpod AI audio recorder.

    **Device API** — recorders push WAV recordings with a resumable upload protocol:

    1. `POST /uploads` with the device's own `recordingId`, the total `size` and the
       `sha256` of the WAV file. Repeating the call with the same `recordingId` returns the
       existing upload (idempotent).
    2. `PATCH /uploads/{uploadId}` with an `Upload-Offset` header and a chunk of bytes as
       body. The whole file may be sent in one PATCH. After a dropped connection,
       `GET /uploads/{uploadId}` returns the offset to resume from.
    3. When the last byte arrives, the checksum and WAV header are verified and the status
       becomes `received`. The recording is then transcoded to FLAC and archived in S3 in
       the background (status `stored`).

    **Pocket** — each user connects their own Pocket recorder (heypocketai.com): its webhook
    (`POST /webhooks/pocket/{webhookId}`, URL from `GET /me/pocket`) announces recordings,
    whose audio is then downloaded with the user's Pocket API key and archived like device
    uploads. WAV and MP3 files can also be uploaded from the browser (`POST /recordings`).

    **AI processing** — every archived recording is transcribed and then summarized through
    OpenRouter, with the API key and models set by an administrator
...

...
    password is `ADMIN_PASSWORD` until one is set in the UI.

...
servers:
...
security: []
...
paths:
...
  /auth/login:
    post:
      tags: ["Web UI sign-in"]
      summary: Sign in to the web UI
      description: Sets the `knowpod_session` cookie. Limited to 10 requests per minute per client IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: Signed in
          headers:
            Set-Cookie: { schema: { type: string, example: knowpod_session=…; Path=/; HttpOnly; SameSite=Strict } }
          content: { application/json: { schema: { $ref: "#/components/schemas/Account" } } }
        "401":
          description: Invalid email or password
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "429": { description: Too many login attempts }
...
  /auth/me:
    get:
      tags: ["Web UI sign-in"]
      summary: The signed-in account
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Signed in
          content: { application/json: { schema: { $ref: "#/components/schemas/Account" } } }
        "401": { $ref: "#/components/responses/Error" }
  /auth/password:
    put:
      tags: ["Web UI sign-in"]
      summary: Change the password
      description: |
        Stores the new password, which from then on replaces the default password from the
        environment. All other sessions of the account are ended.
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, format: password }
                newPassword: { type: string, format: password, minLength: 8, maxLength: 72 }
      responses:
        "204": { description: Password changed }
        "400":
          description: New password too short or too long
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "401": { $ref: "#/components/responses/Error" }
        "403":
          description: Current password is wrong
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /uploads:
    post:
      tags: ["Device uploads"]
      summary: Create (or resume) an upload
      security: [{ deviceToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUpload" }
      responses:
        "201":
          description: Upload created
          headers: { Upload-Offset: { $ref: "#/components/headers/UploadOffset" } }
          content: { application/json: { schema: { $ref: "#/components/schemas/Upload" } } }
        "200":
          description: An upload with this recordingId already exists; its current state
          headers: { Upload-Offset: { $ref: "#/components/headers/UploadOffset" } }
          content: { application/json: { schema: { $ref: "#/components/schemas/Upload" } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "409":
          description: The recordingId exists with a different size or checksum
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "413": { $ref: "#/components/responses/Error" }
  /uploads/{uploadId}:
    parameters:
      - { name: uploadId, in: path, required: true, schema: { type: string } }
    get:
      tags: ["Device uploads"]
      summary: Upload state (offset to resume from)
      security: [{ deviceToken: [] }]
      responses:
        "200":
          description: Current state
          headers: { Upload-Offset: { $ref: "#/components/headers/UploadOffset" } }
          content: { application/json: { schema: { $ref: "#/components/schemas/Upload" } } }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: ["Device uploads"]
      summary: Append a chunk
      description: |
        Appends the body at `Upload-Offset`, which must equal the current offset. Bytes
        received before a dropped connection are kept. Sending a chunk to a completed upload
        is a no-op that returns its final state.
      security: [{ deviceToken: [] }]
      parameters:
        - name: Upload-Offset
          in: header
          required: true
          schema: { type: integer, format: int64, minimum: 0 }
      requestBody:
        required: true
        content:
          application/offset+octet-stream:
            schema: { type: string, format: binary }
      responses:
        "200":
          description: Chunk stored. `status` is `received` once the file is complete and verified.
          headers: { Upload-Offset: { $ref: "#/components/headers/UploadOffset" } }
          content: { application/json: { schema: { $ref: "#/components/schemas/Upload" } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: Offset mismatch; resume from `offset`
          headers: { Upload-Offset: { $ref: "#/components/headers/UploadOffset" } }
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "413":
          description: The chunk would exceed the declared size
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422":
          description: |
            Checksum mismatch (the upload is reset to offset 0; resend) or the file is not a
            supported WAV (the recording is marked `failed`).
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /me/pocket:
    get:
      tags: ["Your account"]
      summary: Your Pocket integration
      description: Your webhook path and whether the secret and API key are set (their values are never returned).
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Settings
          content: { application/json: { schema: { $ref: "#/components/schemas/PocketSettings" } } }
        "401": { $ref: "#/components/responses/Error" }
    put:
      tags: ["Your account"]
      summary: Change your Pocket integration
      description: Fields left out stay unchanged; empty strings remove the value.
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                webhookSecret: { type: string, description: Signing secret shown by Pocket when the webhook is created }
                apiKey: { type: string, example: pk_… }
      responses:
        "200":
          description: Saved
          content: { application/json: { schema: { $ref: "#/components/schemas/PocketSettings" } } }
  /ai/status:
    get:
      tags: ["Your account"]
      summary: Whether AI processing is configured
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                type: object
                properties:
                  transcription: { type: boolean }
                  summary: { type: boolean }
  /ai/models:
    get:
      tags: ["Your account"]
      summary: Models available for transcription and summaries
      description: From OpenRouter's model list. Transcription needs audio input; summaries need text input. Any signed-in user may list them (to pick a summary model per recording).
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      responses:
        "200":
          description: Model options
          content:
            application/json:
              schema:
                type: object
                properties:
                  transcription: { type: array, items: { $ref: "#/components/schemas/ModelOption" } }
                  summary: { type: array, items: { $ref: "#/components/schemas/ModelOption" } }
        "502":
          description: OpenRouter could not be reached
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /ai/languages:
    get:
      tags: ["Your account"]
      summary: Languages a summary can be written in
      description: Language tags accepted as `language` in summary options, besides `auto` (the transcript's language).
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      responses:
        "200":
          description: Language tags
          content:
            application/json:
              schema: { type: array, items: { type: string, example: de-DE } }
  /me/preferences:
    put:
      tags: ["Your account"]
      summary: Change your preferences
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                language: { type: string, enum: ["", en, de], description: Web UI language; empty follows the browser }
      responses:
        "200":
          description: Saved
          content: { application/json: { schema: { $ref: "#/components/schemas/Account" } } }
        "400": { $ref: "#/components/responses/Error" }
  /themes:
    get:
      tags: ["Themes"]
      summary: List summary themes
      description: The built-in themes (`builtIn`, fixed IDs such as `auto`, `meeting`) followed by your own.
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      responses:
        "200":
          description: Themes
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Theme" } }
    post:
      tags: ["Themes"]
      summary: Create a theme
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ThemeInput" } } }
      responses:
        "201":
          description: Created
          content: { application/json: { schema: { $ref: "#/components/schemas/Theme" } } }
        "400": { $ref: "#/components/responses/Error" }
  /themes/{themeId}:
    parameters:
      - { name: themeId, in: path, required: true, schema: { type: string } }
    put:
      tags: ["Themes"]
      summary: Change one of your themes
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ThemeInput" } } }
      responses:
        "200":
          description: Saved
          content: { application/json: { schema: { $ref: "#/components/schemas/Theme" } } }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: ["Themes"]
      summary: Delete one of your themes
      description: Summaries made with it keep their text; regenerating them uses the Auto theme.
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Error" }
  /webhooks/pocket/{webhookId}:
    post:
      tags: ["Integrations"]
      summary: Webhook for Pocket (heypocketai.com)
      description: |
        Each user has their own webhook URL (see `GET /me/pocket`) to configure in their
        Pocket app. Requests are authenticated by their signature: `X-HeyPocket-Signature`
        must be the hex HMAC-SHA256 of `<X-HeyPocket-Timestamp>.<raw body>`, keyed with that
        user's webhook signing secret, and the timestamp (Unix milliseconds) must be within
        5 minutes of the server clock.

        Every event that names a recording queues it once for the user (repeated deliveries
        are ignored); `recording.deleted` is ignored. The audio is then downloaded in the
        background with the user's Pocket API key and archived in S3.
      parameters:
        - { name: webhookId, in: path, required: true, schema: { type: string } }
        - { name: X-HeyPocket-Signature, in: header, required: true, schema: { type: string } }
        - { name: X-HeyPocket-Timestamp, in: header, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event: { type: string, example: recording.created }
                timestamp: { type: string, format: date-time }
                recording:
                  type: object
                  properties:
                    id: { type: string }
                    title: { type: string }
                    duration: { type: number }
                    recordingAt: { type: string, format: date-time }
      responses:
        "200":
          description: Accepted. `status` is `queued`, `duplicate` or `ignored`.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [queued, duplicate, ignored] }
        "400": { $ref: "#/components/responses/Error" }
        "401":
          description: Missing, wrong or expired signature
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "413": { $ref: "#/components/responses/Error" }
        "404":
          description: Unknown webhook URL
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "503":
          description: The user hasn't set both the webhook secret and the API key
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /devices:
    post:
      tags: ["Devices"]
      summary: Register a device for the signed-in user
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 100 }
      responses:
        "201":
          description: Device registered. The token is only returned here.
          content:
            application/json:
              schema:
                type: object
                properties:
                  device: { $ref: "#/components/schemas/Device" }
                  token: { type: string, example: kpd_Zx3... }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
    get:
      tags: ["Devices"]
      summary: List your devices
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      responses:
        "200":
          description: All devices
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Device" } }
  /devices/{deviceId}/token:
    post:
      tags: ["Devices"]
      summary: Issue a new device token
      description: |
        Replaces the device's token, e.g. when the old one was lost. The old token stops
        working immediately. The device keeps its ID and recordings. The new token is only
        returned here.
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      parameters:
        - { name: deviceId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: New token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  device: { $ref: "#/components/schemas/Device" }
                  token: { type: string, example: kpd_Zx3... }
        "401": { $ref: "#/components/responses/Error" }
        "404":
          description: Unknown or removed device
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /devices/{deviceId}:
    delete:
      tags: ["Devices"]
      summary: Remove (revoke) a device
      description: The device's token stops working permanently. Its recordings are kept.
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      parameters:
        - { name: deviceId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Revoked }
        "404": { $ref: "#/components/responses/Error" }
  /recordings:
    get:
      tags: ["Recordings"]
      summary: List your recordings, newest first
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      parameters:
        - { name: deviceId, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/Status" } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
        - name: full
          in: query
          description: Include transcript and summary text (left out by default to keep lists small)
          schema: { type: boolean }
      responses:
        "200":
          description: Recordings
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Recording" } }
    post:
      tags: ["Recordings"]
      summary: Upload an audio file
      description: |
        Uploads a WAV (integer PCM) or MP3 file as the raw request body. It is processed like
        other recordings: WAV is archived as FLAC, MP3 as it is, and both are transcribed
        and summarized. The format is detected from the content, not the file name.
      security: [{ sessionCookie: [] }]
      parameters:
        - name: X-Filename
          in: header
          description: URL-encoded original file name; becomes the title until the summary has one
          schema: { type: string }
        - name: X-Recorded-At
          in: header
          description: When the audio was recorded (RFC 3339), e.g. the file's modification time
          schema: { type: string, format: date-time }
      requestBody:
        required: true
        content:
          audio/wav: { schema: { type: string, format: binary } }
          audio/mpeg: { schema: { type: string, format: binary } }
      responses:
        "201":
          description: Stored; processing starts
          content: { application/json: { schema: { $ref: "#/components/schemas/Recording" } } }
        "400": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "415":
          description: Not a WAV or MP3 file
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422":
          description: A WAV file that isn't integer PCM
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /recordings/{recordingId}:
    delete:
      tags: ["Recordings"]
      summary: Delete a recording
      description: Removes the recording, its transcript and summary, and its audio from storage.
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      parameters:
        - { name: recordingId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Error" }
    get:
      tags: ["Recordings"]
      summary: Get a recording
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      parameters:
        - { name: recordingId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Recording
          content: { application/json: { schema: { $ref: "#/components/schemas/Recording" } } }
        "404": { $ref: "#/components/responses/Error" }
  /recordings/{recordingId}/audio:
    get:
      tags: ["Recordings"]
      summary: Stream or download the archived audio
      description: |
        FLAC for uploaded WAV recordings; the original format (e.g. MP3, M4A) for recordings
        fetched from Pocket. Supports a single `Range` for seeking; `?download=1` sends it as
        an attachment.
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      parameters:
        - { name: recordingId, in: path, required: true, schema: { type: string } }
        - { name: download, in: query, schema: { type: boolean } }
        - { name: Range, in: header, schema: { type: string, example: bytes=0-1023 } }
      responses:
        "206": { description: Partial content for a Range request }
        "416": { description: Range outside the file }
        "200":
          description: Audio
          content:
            audio/flac:
              schema: { type: string, format: binary }
            audio/*:
              schema: { type: string, format: binary }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: Not archived yet
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /recordings/{recordingId}/retranscribe:
    post:
      tags: ["Recordings"]
      summary: Transcribe again
      description: Discards the transcript and summary; the recording is transcribed and summarized anew.
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      parameters:
        - { name: recordingId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Queued
          content: { application/json: { schema: { $ref: "#/components/schemas/Recording" } } }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: The audio has not been archived yet
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /recordings/{recordingId}/resummarize:
    post:
      tags: ["Recordings"]
      summary: Summarize again
      description: |
        Discards the summary and summarizes the existing transcript anew. An optional body
        sets new summary options, which are kept for later regenerations.
      requestBody:
        required: false
        content: { application/json: { schema: { $ref: "#/components/schemas/SummaryOptions" } } }
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      parameters:
        - { name: recordingId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Queued
          content: { application/json: { schema: { $ref: "#/components/schemas/Recording" } } }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: The recording has no transcript yet
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /admin/users:
    get:
      tags: ["Admin: users"]
      summary: List users
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/User" } }
        "403": { $ref: "#/components/responses/Error" }
    post:
      tags: ["Admin: users"]
      summary: Create a user
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password, minLength: 8, maxLength: 72 }
                role: { type: string, enum: [user, admin], default: user }
      responses:
        "201":
          description: Created
          content: { application/json: { schema: { $ref: "#/components/schemas/User" } } }
        "400": { $ref: "#/components/responses/Error" }
        "409":
          description: Email already in use
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /admin/users/{userId}:
    parameters:
      - { name: userId, in: path, required: true, schema: { type: string } }
    put:
      tags: ["Admin: users"]
      summary: Change a user's email or role
      description: The built-in admin (ADMIN_EMAIL) keeps its email and role; the last admin can't be demoted.
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                role: { type: string, enum: [user, admin] }
      responses:
        "200":
          description: Saved
          content: { application/json: { schema: { $ref: "#/components/schemas/User" } } }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      tags: ["Admin: users"]
      summary: Delete a user and their data
      description: |
        Removes the user with their devices, recordings (including the audio in storage)
        and sessions. You can't delete yourself, the built-in admin or the last admin.
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      responses:
        "204": { description: Deleted }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /admin/users/{userId}/password:
    put:
      tags: ["Admin: users"]
      summary: Set a user's password
      description: The user is signed out everywhere. For the built-in admin this replaces ADMIN_PASSWORD.
      security: [{ sessionCookie: [] }, { adminToken: [] }]
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, format: password, minLength: 8, maxLength: 72 }
      responses:
        "204": { description: Password set }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      description: Device token returned by POST /admin/devices
    sessionCookie:
      type: apiKey
      in: cookie
      name: knowpod_session
      description: Web UI session, set by POST /auth/login
    adminToken:
      type: http
      scheme: bearer
      description: The ADMIN_TOKEN environment variable (for scripts)
...
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...